# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建直播房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "创建直播房间",
                "operationId": "create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除房间",
                "operationId": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/get": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取直播房间详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取直播房间详情",
                "operationId": "get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
...
        "/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取直播房间列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取直播房间列表",
                "operationId": "list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RoomInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
...
        "/pause": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "暂停直播",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "暂停直播",
                "operationId": "pause",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LiveStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "开始直播",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "开始直播",
                "operationId": "start",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LiveStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/stop": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "停止直播",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "停止直播",
                "operationId": "stop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LiveStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取TOKEN",
                "operationId": "token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/handler.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新房间详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新房间详情",
                "operationId": "update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateRequest": {
            "type": "object",
            "required": [
                "anchor",
                "title"
            ],
            "properties": {
                "anchor": {
                    "description": "主播id",
                    "type": "string",
                    "example": "主播userId"
                },
                "anchor_nick": {
                    "description": "主播Nick",
                    "type": "string",
                    "example": "主播nick"
                },
                "extends": {
                    "description": "扩展字段，通常是JSON格式字符串",
                    "type": "string",
                    "example": "扩展字段，通常是JSON格式字符串"
                },
                "mode": {
                    "description": "模式，默认0 普通直播，1 连麦直播",
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "notice": {
                    "description": "直播公告",
                    "type": "string",
                    "example": "直播公告"
                },
                "title": {
                    "description": "直播标题",
                    "type": "string",
                    "example": "直播标题"
                }
            }
        },
        "handler.DeleteRequest": {
            "type": "object",
            "required": [
                "id",
                "user_id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "uuid，直播房间id"
                },
                "user_id": {
                    "type": "string",
                    "example": "当前用户id"
                }
            }
        },
...
        "handler.GetRequest": {
            "type": "object",
            "required": [
                "id",
                "user_id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "uuid，直播房间id"
                },
                "user_id": {
                    "type": "string",
                    "example": "当前用户id"
                }
            }
        },
        "handler.ListRequest": {
            "type": "object",
            "required": [
                "page_num",
                "page_size",
                "user_id"
            ],
            "properties": {
                "page_num": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "handler.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "models.RoomInfo": {
            "type": "object",
            "properties": {
                "anchor_id": {
                    "description": "主播Id",
                    "type": "string"
                },
                "anchor_nick": {
                    "description": "主播Nick",
                    "type": "string"
                },
                "chat_id": {
                    "description": "群组Id",
                    "type": "string"
                },
                "cover_url": {
                    "description": "直播封面",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "extends": {
                    "description": "扩展字段",
                    "type": "string"
                },
                "id": {
                    "description": "直播Id",
                    "type": "string"
                },
                "link_info": {
                    "description": "连麦PK 等信息， 动态生成",
                    "$ref": "#/definitions/models.LinkInfo"
                },
                "meeting_id": {
                    "description": "连麦Id",
                    "type": "string"
                },
                "meeting_info": {
                    "description": "连麦成员信息（json序列化）",
                    "type": "string"
                },
                "metrics": {
                    "description": "直播间统计",
                    "$ref": "#/definitions/models.Metrics"
                },
                "mode": {
                    "description": "直播模式 0-普通直播, 1-连麦直播，2-PK直播",
                    "type": "integer"
                },
                "notice": {
                    "description": "直播公告",
                    "type": "string"
                },
                "pull_url_info": {
                    "description": "拉流相关地址信息，动态生成",
                    "$ref": "#/definitions/live.PullLiveInfo"
                },
                "push_url_info": {
                    "description": "推流相关地址信息，动态生成",
                    "$ref": "#/definitions/models.PushLiveInfo"
                },
                "status": {
                    "description": "直播状态，0-准备中，1-已开始，2-已结束",
                    "type": "integer"
                },
                "title": {
                    "description": "直播标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_status": {
                    "description": "用户状态",
                    "$ref": "#/definitions/models.UserStatus"
                },
                "vod_id": {
                    "description": "点播Id",
                    "type": "string"
                },
                "vod_info": {
                    "description": "直播转录制，点播信息",
                    "$ref": "#/definitions/models.VodInfo"
                }
            }
        },
        "models.Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}